Added configuration to use Neopixels with the basic CR-10 (S4 and S5 included)
The Pin-27 board can now be used as LED driver for Neopixels (WS2812B). The Melzi board does not have enough memory to have the LED menu in the LCD, however we can use the  Gcode M150 to change the color of the lights. Event LED is enabled, it switches from white to blue to violet as the bed heats up, then to red as extruder heats up and finally to white as it prints.